+++ src/main/java/com/example/demo/dto/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.example.demo.dataObject.dto;
+package com.example.demo.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
