+++ src/main/java/com/example/demo/dto/SigninDTO.java
@@ -1,11 +1,11 @@
-package com.example.demo.dataObject.vo;
+package com.example.demo.dto;
 
 import lombok.Setter;
 import lombok.Getter;
 
 @Getter
 @Setter
-public class SigninVO {
+public class SigninDTO {
     private String email;
     private String password;
 }
